[backend] Create backup only if the number of records have changed

+++ code/backend/database_backups/backup_from_server.py
@@ -6,14 +6,29 @@ import sys
 
 numberOfRecords = requests.get("http://weatheranalytics.tk:8080/count").json()["count"]
 
-keyFilePath = "~/.ssh/co300.pem"
-if len(sys.argv) > 1:
-    keyFilePath = sys.argv[1]
+isAlreadyBackedUp = False
 
-os.system(
-    f"ssh -i {keyFilePath} Nipun@weatheranalytics.tk 'rm -rf dump;mongodump; rm dump.7z ;7z a dump.7z dump/'"
-)
-currentDatetime = datetime.datetime.now().isoformat().replace(":", "-")
-os.system(
-    f"scp -i {keyFilePath} Nipun@weatheranalytics.tk:/home/Nipun/dump.7z ./{currentDatetime}-{numberOfRecords}.7z"
-)
+filesInCurrrentDirectory = os.listdir()
+for eachFile in filesInCurrrentDirectory:
+    if ".7z" not in eachFile:
+        continue
+
+    numberOfRecordsBackedUpInThisBackup = (
+        eachFile.split("T")[1].split("-")[3].split(".")[0]
+    )
+    if numberOfRecordsBackedUpInThisBackup == numberOfRecords:
+        isAlreadyBackedUp = True
+        break
+
+if not isAlreadyBackedUp:
+    keyFilePath = "~/.ssh/co300.pem"
+    if len(sys.argv) > 1:
+        keyFilePath = sys.argv[1]
+
+    os.system(
+        f"ssh -i {keyFilePath} Nipun@weatheranalytics.tk 'rm -rf dump;mongodump; rm dump.7z ;7z a dump.7z dump/'"
+    )
+    currentDatetime = datetime.datetime.now().isoformat().replace(":", "-")
+    os.system(
+        f"scp -i {keyFilePath} Nipun@weatheranalytics.tk:/home/Nipun/dump.7z ./{currentDatetime}-{numberOfRecords}.7z"
+    )
